Change metadata system to fix broken links
The metadata system ('include fields.md') copied over from the user
guide does not appear to work properly on the main website (most likely
due to the newer Jekyll version in use here). This is causing broken
links on the pages that use them - the quickstart guides. This commit
changes to using a Jekyll plugin to provide the metadata which works
more reliably (and is easier-to-understand code than trying to do it in
Liquid template language) and converts the quickstart guide to use this
new metadata. This fixes the broken links on the quickstart guides.

It also renames quickstart/my-web-cluster.yaml to be
_my-web-cluster.yaml to prevent the file being published; it contains
template variable substitutions so it's not suitable for direct use.